refactor: Update styles for DropdownMenu, Strip, Navbar, and Header components
