+++ src/components/DropdownMenu/DropdownMenuRoot.tsx
@@ -9,7 +9,7 @@ export function DropdownMenuRoot({ children }: { children: React.ReactNode }) {
     return (
         <DropdownMenu.Root onOpenChange={setIsOpen}>
             <DropdownMenu.Trigger className={styles.dropdownMenu}>
-                <Icon.Menu />
+                    {isOpen ? <Icon.Close /> : <Icon.Menu />}
             </DropdownMenu.Trigger>
             <DropdownMenu.Portal>
                 <DropdownMenu.Content
