+++ src/pages/programacao/index.tsx
@@ -4,7 +4,7 @@ import { Header } from "../../shared/header";
 import { ActivityCard } from "../../components/programacao/ActivityCard";
 import { Ticket } from "../../shared/Ticket";
 import triangles from "../../assets/icon/TRIANGLES.svg";
-import activitiesJson from "../../data/activitiesNovo.json";
+import activitiesJson from "../../data/activities.json";
 
 const activities = activitiesJson;
