+++ src/shared/Navbar/NavRoot.tsx
@@ -5,7 +5,6 @@ import { Button } from "../../components/Button";
 import { DropdownMenu } from "../../components/DropdownMenu";
 import Dropdown from "../../components/Dropdown";
 import { Link } from "react-router-dom";
-import React from "react";
 
 interface NavRootProps {
   children: ReactNode;
@@ -16,12 +15,14 @@ export function NavRoot({ children }: NavRootProps) {
     window.open('https://www.even3.com.br/sescomp-vii/', '_blank');
   };
 
-  const [isOpen, setIsOpen] = React.useState(false);
-
+  const handleLogoClick = () => {
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  };
+  
   return (
     <nav className={styles.navbar}>
       <div className={styles.logo}>
-        <Link to="/">
+        <Link to="/" onClick={handleLogoClick}>
           <Icon.Sescomp />
         </Link>
       </div>
@@ -30,9 +31,11 @@ export function NavRoot({ children }: NavRootProps) {
           {children}
           <Dropdown />
         </ul>
-        <Button.Root>
-          <Button.Content label="Participar do Evento" onClick={handleNavigation}></Button.Content>
-        </Button.Root>
+        <div className={styles.participarButton} onClick={handleNavigation}>
+          <Button.Root>
+            <Button.Content label="Participar do Evento"></Button.Content>
+          </Button.Root>
+        </div>
       </div>
       <div className={styles.menu}>
         <DropdownMenu.Root>
@@ -42,13 +45,14 @@ export function NavRoot({ children }: NavRootProps) {
             </ul>
           </DropdownMenu.Item>
           <DropdownMenu.Item>
-            <Button.Root>
-              <Button.Content label="Participar do Evento" onClick={handleNavigation}></Button.Content>
-            </Button.Root>
+            <div className={styles.participarButton} onClick={handleNavigation}>
+              <Button.Root>
+                <Button.Content label="Participar do Evento"></Button.Content>
+              </Button.Root>
+            </div>
           </DropdownMenu.Item>
         </DropdownMenu.Root>
       </div>
-      {isOpen && <div className={styles.overlay} onClick={() => setIsOpen(false)}></div>}
     </nav>
   );
 }
