+++ src/shared/header/styles.module.scss
@@ -4,6 +4,16 @@
     background-position: center;
     width: 100%;
     height: 384px;
+    filter: brightness(1.3) contrast(1.4);
+}
+
+.overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    opacity: 0.4;
 }
 
 .titulo {
@@ -16,7 +26,7 @@
 }
 
 .titulo h1 {
-    font-size: 3.5rem;
+    font-size: 56px;
     font-weight: 700;
     margin-bottom: 0;
     transform: translate(0, +160%) rotate(-5deg);
@@ -29,6 +39,22 @@
 
 @media screen and (max-width: 768px) {
     .titulo {
+        
         transform: scale(0.8) translate(-60%, -70%);
+    }
+    .titulo h1 {
+        transform: translate(0, +180%) rotate(-5deg);
+        font-size: 50px;
+        
+    }
+}
+
+@media screen and (max-width: 576px) {
+    .header {
+        height: 306px;
+    }
+
+    .titulo {
+        transform: scale(0.6) translate(-80%, -80%);
     }
 }
